itmatepk/pyproject.toml
[project]
name = "itmatepk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.4",
    "python-decouple>=3.8",
    "rich>=14.1.0",
]
